pyproject.toml
[project]
name = "qpick"
version = "0.1.0"
description = "django semester project"
authors = [
    {name = "dvodnenko",email = "thekovalskii@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "django-cloudinary-storage (>=0.3.0,<0.4.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-rest-swagger (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
